backend/requirements_complex.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP & CORS
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.0
requests==2.31.0
python-dotenv==1.0.0

# Geolocation for assignment algorithm
geopy==2.4.1

# Utilities
email-validator==2.1.0
structlog==23.2.0
redis==5.0.1
python-json-logger==2.0.7